add phpstan checks

+++ uninstall.php
@@ -20,6 +20,12 @@ delete_transient( 'next_meetup_hint' );
 
 // delete settings on user profiles.
 foreach ( get_users() as $user ) {
+	// bail if user is not a user object.
+	if( ! $user instanceof WP_User ) {
+		continue;
+	}
+
+	// delete the settings on this user.
 	delete_user_meta( $user->ID, 'hide_next_meet_hint' );
 	delete_user_meta( $user->ID, 'next_meet_hint_days' );
 	delete_user_meta( $user->ID, 'hide_next_meetup_hint_for_2_days' );
